requirements.txt
requests            
beautifulsoup4      
selenium            
google-auth         
streamlit           
python-dotenv
google-generativeai